Add files via upload
SSL and Tls: Theory and Practice  SSL (secure socket layer) and TLS (Transport Layer Security) are widely deployed security protocols that are used in all kinds of web-based e-commerce and e-business applications. This book offers an introduction to these protocols. It shows how to employ SSL and TLS and configure security solutions that are based on the use of SSL/TLS protocols